django-backend/ecommerce_webapp/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
django-cors-headers
django-mptt
django-filter
Pillow
